Clean-up in the node.
